requirements.txt
dataclasses==0.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.28
django-import-export==2.5.0
django-stubs==0.10.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mypy==0.910
mypy-extensions==0.4.4
odfpy==1.4.1
openpyxl==2.6.4
psycopg2==2.8.6
pyngrok==6.0.0
pytz==2023.3
PyYAML==5.3.1
sqlparse==0.4.4
tablib==2.0.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
xlrd==1.2.0
xlwt==1.3.0
